Move implementation of binding and unbinding into backstore
I believe we achieve better encapsulation that way.
